fixed: for eager fetch hibernate can make duplicates if a list is used
When a 'körvezető' edited his own posts for a his membership it
would lead to an exception, because his posts were duplicated. It tried
to remove the 'körvezető' post as well, which is not possible.

Switched List<> to Set<>. Order of posts does not matter anyway.

related stackoverflow question:
http://stackoverflow.com/a/1093694